backend_flask/requirements.txt
Flask==2.3.2
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
# pydantic optional if you want typed models later
# pydantic==1.10.11
